llvm_mode has 32-bit __afl_prev_loc, gcc_mode is synced now.

+++ gcc_plugin/afl-gcc-rt.o.c
@@ -41,7 +41,7 @@
 
 u8  __afl_area_initial[MAP_SIZE];
 u8* __afl_area_ptr = __afl_area_initial;
-u16 __afl_prev_loc;
+u32 __afl_prev_loc;
 
 
 /* Running in persistent mode? */
